+++ sandpack-react/src/components/Tests/style.ts
@@ -1,11 +1,17 @@
 import { css } from "../../styles";
 
+export const setTestTheme = (isDark: boolean): Record<string, string> => ({
+  "--test-pass": isDark ? "#18df16" : "#15c213",
+  "--test-fail": isDark ? "#df162b" : "#c21325",
+  "--test-skip": isDark ? "#eace2b" : "#c2a813",
+});
+
 const color = css({
   variants: {
     status: {
-      pass: { color: "$colors$accent" },
-      fail: { color: "$colors$error" },
-      skip: { color: "$colors$warning" },
+      pass: { color: "var(--test-pass)" },
+      fail: { color: "var(--test-fail)" },
+      skip: { color: "var(--test-skip)" },
     },
   },
 });
@@ -17,8 +23,8 @@ export const skipTextClassName = color({ status: "skip" });
 const background = css({
   variants: {
     status: {
-      pass: { background: "$colors$accent" },
-      fail: { background: "$colors$errorSurface" },
+      pass: { background: "var(--test-pass)", color: "$colors$surface1" },
+      fail: { background: "var(--test-fail)", color: "$colors$surface1" },
     },
   },
 });
